+++ src/ui/modals/info_list_modal.rs
@@ -15,7 +15,7 @@ use ratatui::{
 use super::{Modal, RectExt};
 use crate::{
     config::keys::CommonAction,
-    context::Ctx,
+    ctx::Ctx,
     mpd::commands::Song,
     shared::{
         ext::duration::DurationExt,
