+++ src/ui/modals/info_modal.rs
@@ -16,7 +16,7 @@ use ratatui::{
 use super::{BUTTON_GROUP_SYMBOLS, Modal, RectExt};
 use crate::{
     config::{Size, keys::CommonAction},
-    context::Ctx,
+    ctx::Ctx,
     shared::{
         key_event::KeyEvent,
         macros::pop_modal,
