+++ src/ui/panes/albums.rs
@@ -6,7 +6,7 @@ use super::Pane;
 use crate::{
     MpdQueryResult,
     config::{keys::actions::Position, sort_mode::SortOptions, tabs::PaneType},
-    context::Ctx,
+    ctx::Ctx,
     mpd::{
         client::Client,
         commands::Song,
